Changed test directory structure to match main Changed CHAR from 元 to its unicode \u5143
