+++ src/main/java/de/espend/idea/php/phpunit/type/MockProphecyTypeProvider.java
@@ -16,7 +16,7 @@ import org.jetbrains.annotations.Nullable;
 import java.util.*;
 
 public class MockProphecyTypeProvider implements PhpTypeProvider4 {
-    public static final char CHAR = '元';
+    public static final char CHAR = '\u5143';
     public static char TRIM_KEY = '\u0192';
 
     private static final Collection<String> PHPUNIT_CLASSES = new HashSet<String>() {{
