+++ src/test/java/de/espend/idea/php/phpunit/completion/fixtures/PhpUnitCompletionContributor.php
@@ -1,5 +1,15 @@
 <?php
 
+namespace Foo
+{
+    class Bar
+    {
+        public function getFoobar()
+        {
+        }
+    }
+}
+
 namespace PHPUnit\Framework
 {
     abstract class TestCase
@@ -12,13 +22,6 @@ namespace PHPUnit\Framework
         protected function createMock($originalClassName)
         {
             return new \PHPUnit_Framework_MockObject_MockObject();
-        }
-
-        /**
-         * @return \PHPUnit\Framework\MockObject\MockBuilder
-         */
-        public function getMockBuilder() {
-            return new \PHPUnit\Framework\MockObject\MockBuilder();
         }
     }
 }
@@ -38,6 +41,9 @@ namespace
         }
     }
 
+    /**
+     * @method PHPUnit_Framework_MockObject_Builder_InvocationMocker method($constraint)
+     */
     class PHPUnit_Framework_MockObject_MockObject
     {
         /**
@@ -51,41 +57,19 @@ namespace
 
     class PHPUnit_Framework_MockObject_MockBuilder
     {
+
     }
 
     class PHPUnit_Framework_MockObject_Builder_InvocationMocker
     {
-    }
-}
-
-namespace PHPUnit\Framework\MockObject
-{
-    class MockBuilder
-    {
         /**
-         * @return $this
+         * @param \PHPUnit_Framework_Constraint|string $constraint
+         *
+         * @return \PHPUnit_Framework_MockObject_Builder_InvocationMocker
          */
-        public function disableOriginalConstructor()
+        public function method($constraint)
         {
-            return $this;
+            return new PHPUnit_Framework_MockObject_Builder_InvocationMocker();
         }
-
-        /**
-         * @return $this
-         */
-        public function getMock()
-        {
-            return $this;
-        }
-    }
-}
-
-namespace Foo
-{
-    class Bar
-    {
-       public function getFooBar()
-       {
-       }
     }
 }
