+++ src/test/java/de/espend/idea/php/phpunit/intention/MethodExceptionIntentionActionTest.java
@@ -1,7 +1,7 @@
-package de.espend.idea.php.phpunit.tests.intention;
+package de.espend.idea.php.phpunit.intention;
 
 import com.jetbrains.php.lang.PhpFileType;
-import de.espend.idea.php.phpunit.tests.PhpUnitLightCodeInsightFixtureTestCase;
+import de.espend.idea.php.phpunit.PhpUnitLightCodeInsightFixtureTestCase;
 
 /**
  * @author Daniel Espendiller <daniel@espendiller.net>
@@ -15,7 +15,7 @@ public class MethodExceptionIntentionActionTest extends PhpUnitLightCodeInsightF
     }
 
     public String getTestDataPath() {
-        return "src/test/java/de/espend/idea/php/phpunit/tests/intention/fixtures";
+        return "src/test/java/de/espend/idea/php/phpunit/intention/fixtures";
     }
 
     public void testThatIntentionIsAvailableForConstructorContext() {
