+++ src/test/java/de/espend/idea/php/phpunit/intention/fixtures/ConstructorMockIntention.php
