+++ src/test/java/de/espend/idea/php/phpunit/type/fixtures/RevealProphecyTypeProvider.php
